+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## StreamChat
 ### ✅ Added
 - `VideoAttachmentPayload` now has `originalWidth`, `originalHeight`, and `duration`; passed from `localMetadata` when creating payloads from local files (for custom CDN) [#3991](https://github.com/GetStream/stream-chat-swift/pull/3991)
+- Add freeze and unfreeze channel support to `LivestreamChannelController` [#3996](https://github.com/GetStream/stream-chat-swift/pull/3996)
 
 ## StreamChatUI
 ### ✅ Added